feat(ui): admin edit shortcut on MCP pill, restructured edit modal
Surface the MCP server edit modal directly from the pill (admin-gated
pencil icon), and reorganize the modal so connection essentials are
visible by default with the OAuth long-tail collapsed under "Advanced".

UI:
- MCPServerPill: small pencil icon at end of tag content for admins
  (uses existing usePermissions().isAdmin); click stops propagation so
  it doesn't trigger OAuth/refresh, opens the new edit modal inline.
- Edit modal: Basic + Connection both open by default (was: Basic
  only). Connection now flows key fields → action buttons (Test Auth /
  Start OAuth / Disconnect / Test Connection) → results → env vars →
  collapsed "Advanced — OAuth endpoints" with a dot on the header when
  any auto-discoverable endpoint has been customized.

Refactor (no behavior change):
- Extract MCPServerFormFields out of MCPServersTable.tsx.
- New self-contained MCPServerEditModal owns its form/state/handlers
  so MCPServerPill can mount it without lifting state.
- MCPServersTable shrinks 1656 → ~510 lines; create + view modals
  remain inline.
- Form hydration in the edit modal is keyed on mcp_server_id so
  real-time WebSocket updates to the open server don't clobber
  in-progress edits (matches original button-driven hydration).
