+++ apps/agor-ui/src/components/MCPServerEditModal.tsx
@@ -1,0 +1,298 @@
+import type { AgorClient, MCPServer, UpdateMCPServerInput } from '@agor-live/client';
+import { Form, Modal } from 'antd';
+import { useEffect, useState } from 'react';
+import { useThemedMessage } from '@/utils/message';
+import { MCPServerFormFields } from './SettingsModal/MCPServerFormFields';
+import { extractOAuthConfig } from './SettingsModal/mcp-oauth-utils';
+
+export interface MCPServerEditModalProps {
+  /** The server being edited. Modal opens when this is non-null and `open` is true. */
+  server: MCPServer | null;
+  open: boolean;
+  client: AgorClient | null;
+  onClose: () => void;
+  /** Optional parent state-management callback (mirrors what MCPServersTable used to do). */
+  onUpdate?: (serverId: string, updates: UpdateMCPServerInput) => void;
+}
+
+interface TestResult {
+  success: boolean;
+  toolCount: number;
+  resourceCount: number;
+  promptCount: number;
+  error?: string;
+  tools?: Array<{ name: string; description: string }>;
+  resources?: Array<{ name: string; uri: string; mimeType?: string }>;
+  prompts?: Array<{ name: string; description: string }>;
+}
+
+/**
+ * Self-contained "Edit MCP Server" modal.
+ *
+ * Hydrates its own form from `server`, owns transport/authType/test state,
+ * and persists updates via the `mcp-servers` Feathers service. Used by
+ * both `MCPServersTable` (settings) and `MCPServerPill` (admin shortcut).
+ */
+export const MCPServerEditModal: React.FC<MCPServerEditModalProps> = ({
+  server,
+  open,
+  client,
+  onClose,
+  onUpdate,
+}) => {
+  const { showSuccess, showError } = useThemedMessage();
+  const [form] = Form.useForm();
+  const [transport, setTransport] = useState<'stdio' | 'http' | 'sse'>('stdio');
+  const [authType, setAuthType] = useState<'none' | 'bearer' | 'jwt' | 'oauth'>('none');
+  const [testing, setTesting] = useState(false);
+  const [testResult, setTestResult] = useState<TestResult | null>(null);
+
+  // Hydrate the form when the modal opens or the user swaps to a different
+  // server. Intentionally NOT keyed on `server` itself — that would clobber
+  // in-progress edits whenever the parent's WebSocket sync re-emits the
+  // record.
+  // biome-ignore lint/correctness/useExhaustiveDependencies: see comment above
+  useEffect(() => {
+    if (!open || !server) return;
+
+    setTestResult(null);
+    const serverAuthType = (server.auth?.type as 'none' | 'bearer' | 'jwt' | 'oauth') || 'none';
+    setAuthType(serverAuthType);
+    setTransport(server.transport || (server.url ? 'http' : 'stdio'));
+
+    // Reset first to clear any stale fields registered for a different auth type.
+    form.resetFields();
+
+    const formValues: Record<string, unknown> = {
+      name: server.name,
+      display_name: server.display_name,
+      description: server.description,
+      transport: server.transport || (server.url ? 'http' : 'stdio'),
+      command: server.command,
+      args: server.args?.join(', '),
+      url: server.url,
+      scope: server.scope,
+      enabled: server.enabled,
+      env: server.env ? JSON.stringify(server.env, null, 2) : undefined,
+      auth_type: serverAuthType,
+      tool_permissions: server.tool_permissions || {},
+    };
+
+    // Only set fields for the active auth type to avoid AntD validating hidden fields.
+    if (serverAuthType === 'bearer') {
+      formValues.auth_token = server.auth?.token;
+    } else if (serverAuthType === 'jwt') {
+      formValues.jwt_api_url = server.auth?.api_url;
+      formValues.jwt_api_token = server.auth?.api_token;
+      formValues.jwt_api_secret = server.auth?.api_secret;
+    } else if (serverAuthType === 'oauth') {
+      formValues.oauth_authorization_url = server.auth?.oauth_authorization_url;
+      formValues.oauth_token_url = server.auth?.oauth_token_url;
+      formValues.oauth_client_id = server.auth?.oauth_client_id;
+      formValues.oauth_client_secret = server.auth?.oauth_client_secret;
+      formValues.oauth_scope = server.auth?.oauth_scope;
+      formValues.oauth_grant_type = server.auth?.oauth_grant_type || 'client_credentials';
+      formValues.oauth_mode = server.auth?.oauth_mode || 'per_user';
+    }
+
+    form.setFieldsValue(formValues);
+  }, [open, server?.mcp_server_id, form]);
+
+  const closeAndReset = () => {
+    form.resetFields();
+    setTransport('stdio');
+    setAuthType('none');
+    setTestResult(null);
+    onClose();
+  };
+
+  const handleTestConnection = async () => {
+    if (!client || !server) {
+      showError('Client not available');
+      return;
+    }
+
+    const values = form.getFieldsValue(true);
+
+    if (!values.url) {
+      showError('URL is required to test connection');
+      return;
+    }
+    if (values.transport === 'stdio') {
+      showError('Connection test is not available for stdio transport');
+      return;
+    }
+
+    setTesting(true);
+    setTestResult(null);
+
+    try {
+      let auth:
+        | {
+            type: 'none' | 'bearer' | 'jwt' | 'oauth';
+            token?: string;
+            api_url?: string;
+            api_token?: string;
+            api_secret?: string;
+            oauth_token_url?: string;
+            oauth_client_id?: string;
+            oauth_client_secret?: string;
+            oauth_scope?: string;
+            oauth_grant_type?: string;
+            oauth_mode?: 'per_user' | 'shared';
+          }
+        | undefined;
+
+      if (values.auth_type && values.auth_type !== 'none') {
+        auth = { type: values.auth_type };
+        if (values.auth_type === 'bearer') {
+          auth.token = values.auth_token;
+        } else if (values.auth_type === 'jwt') {
+          auth.api_url = values.jwt_api_url;
+          auth.api_token = values.jwt_api_token;
+          auth.api_secret = values.jwt_api_secret;
+        } else if (values.auth_type === 'oauth') {
+          const oauthConfig = extractOAuthConfig(values);
+          Object.assign(auth, oauthConfig);
+        }
+      }
+
+      const data = (await client.service('mcp-servers/discover').create({
+        mcp_server_id: server.mcp_server_id,
+        url: values.url,
+        transport: values.transport || 'http',
+        auth,
+      })) as {
+        success: boolean;
+        error?: string;
+        capabilities?: { tools: number; resources: number; prompts: number };
+        tools?: Array<{ name: string; description: string }>;
+        resources?: Array<{ name: string; uri: string; mimeType?: string }>;
+        prompts?: Array<{ name: string; description: string }>;
+      };
+
+      if (data.success && data.capabilities) {
+        const result: TestResult = {
+          success: true,
+          toolCount: data.capabilities.tools,
+          resourceCount: data.capabilities.resources,
+          promptCount: data.capabilities.prompts,
+          tools: data.tools,
+          resources: data.resources,
+          prompts: data.prompts,
+        };
+        setTestResult(result);
+        showSuccess(
+          `Connection successful: ${result.toolCount} tools, ${result.resourceCount} resources, ${result.promptCount} prompts`
+        );
+      } else {
+        const errorMsg = data.error || 'Connection test failed';
+        setTestResult({
+          success: false,
+          toolCount: 0,
+          resourceCount: 0,
+          promptCount: 0,
+          error: errorMsg,
+        });
+        showError(errorMsg);
+      }
+    } catch (error) {
+      const errorMessage = error instanceof Error ? error.message : 'Unknown error';
+      setTestResult({
+        success: false,
+        toolCount: 0,
+        resourceCount: 0,
+        promptCount: 0,
+        error: errorMessage,
+      });
+      showError(`Connection test failed: ${errorMessage}`);
+    } finally {
+      setTesting(false);
+    }
+  };
+
+  const handleSave = async () => {
+    if (!server || !client) return;
+
+    try {
+      await form.validateFields();
+      const values = form.getFieldsValue(true);
+
+      const updates: UpdateMCPServerInput = {
+        display_name: values.display_name,
+        description: values.description,
+        scope: values.scope,
+        enabled: values.enabled,
+        transport: values.transport,
+      };
+
+      if (values.transport === 'stdio') {
+        updates.command = values.command;
+        updates.args = values.args?.split(',').map((arg: string) => arg.trim()) || [];
+      } else {
+        updates.url = values.url;
+      }
+
+      if (values.env) {
+        try {
+          updates.env = JSON.parse(values.env);
+        } catch {
+          // Invalid JSON, skip
+        }
+      }
+
+      if (values.auth_type && values.auth_type !== 'none') {
+        updates.auth = { type: values.auth_type };
+        if (values.auth_type === 'bearer') {
+          updates.auth.token = values.auth_token;
+        } else if (values.auth_type === 'jwt') {
+          updates.auth.api_url = values.jwt_api_url;
+          updates.auth.api_token = values.jwt_api_token;
+          updates.auth.api_secret = values.jwt_api_secret;
+        } else if (values.auth_type === 'oauth') {
+          const oauthConfig = extractOAuthConfig(values);
+          Object.assign(updates.auth, oauthConfig);
+        }
+      } else {
+        updates.auth = undefined;
+      }
+
+      await client.service('mcp-servers').patch(server.mcp_server_id, updates);
+      onUpdate?.(server.mcp_server_id, updates);
+
+      showSuccess('MCP server updated successfully');
+      closeAndReset();
+    } catch (error) {
+      const errorMessage = error instanceof Error ? error.message : 'Failed to update server';
+      showError(errorMessage);
+    }
+  };
+
+  return (
+    <Modal
+      title="Edit MCP Server"
+      open={open}
+      onOk={handleSave}
+      onCancel={closeAndReset}
+      okText="Save"
+      width={600}
+      destroyOnHidden
+    >
+      <Form form={form} layout="vertical" style={{ marginTop: 16 }}>
+        <MCPServerFormFields
+          mode="edit"
+          transport={transport}
+          onTransportChange={setTransport}
+          authType={authType}
+          onAuthTypeChange={setAuthType}
+          form={form}
+          client={client}
+          serverId={server?.mcp_server_id}
+          onTestConnection={handleTestConnection}
+          testing={testing}
+          testResult={testResult}
+        />
+      </Form>
+    </Modal>
+  );
+};
