+++ apps/agor-ui/src/components/SettingsModal/MCPServersTable.tsx
@@ -1,37 +1,23 @@
 import type { CreateMCPServerInput, MCPServer, UpdateMCPServerInput } from '@agor-live/client';
+import { DeleteOutlined, EditOutlined, EyeOutlined, PlusOutlined } from '@ant-design/icons';
 import {
-  ApiOutlined,
-  DeleteOutlined,
-  DownOutlined,
-  EditOutlined,
-  EyeOutlined,
-  PlusOutlined,
-} from '@ant-design/icons';
-import type { FormInstance } from 'antd';
-import {
-  Alert,
   Badge,
   Button,
-  Collapse,
   Descriptions,
   Form,
-  Input,
   Modal,
   Popconfirm,
-  Select,
   Space,
-  Switch,
   Table,
   Tag,
-  Tooltip,
   Typography,
 } from 'antd';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { mapToSortedArray } from '@/utils/mapHelpers';
 import { useThemedMessage } from '@/utils/message';
-import { extractOAuthConfig, extractOAuthConfigForTesting } from './mcp-oauth-utils';
-
-const { TextArea } = Input;
+import { MCPServerEditModal } from '../MCPServerEditModal';
+import { MCPServerFormFields } from './MCPServerFormFields';
+import { extractOAuthConfig } from './mcp-oauth-utils';
 
 interface MCPServersTableProps {
   mcpServerById: Map<string, MCPServer>;
@@ -41,856 +27,16 @@ interface MCPServersTableProps {
   onDelete?: (serverId: string) => void;
 }
 
-interface MCPServerFormFieldsProps {
-  mode: 'create' | 'edit';
-  transport?: 'stdio' | 'http' | 'sse';
-  onTransportChange?: (transport: 'stdio' | 'http' | 'sse') => void;
-  authType?: 'none' | 'bearer' | 'jwt' | 'oauth';
-  onAuthTypeChange?: (authType: 'none' | 'bearer' | 'jwt' | 'oauth') => void;
-  form: FormInstance;
-  client: import('@agor-live/client').AgorClient | null;
-  serverId?: string;
-  onTestConnection?: () => Promise<void>;
-  testing?: boolean;
-  testResult?: {
-    success: boolean;
-    toolCount: number;
-    resourceCount: number;
-    promptCount: number;
-    error?: string;
-    tools?: Array<{ name: string; description: string }>;
-    resources?: Array<{ name: string; uri: string; mimeType?: string }>;
-    prompts?: Array<{ name: string; description: string }>;
-  } | null;
-  /** Callback to save server first before OAuth flow (for new servers) */
-  onSaveFirst?: () => Promise<string | null>;
+interface TestResult {
+  success: boolean;
+  toolCount: number;
+  resourceCount: number;
+  promptCount: number;
+  error?: string;
+  tools?: Array<{ name: string; description: string }>;
+  resources?: Array<{ name: string; uri: string; mimeType?: string }>;
+  prompts?: Array<{ name: string; description: string }>;
 }
-
-const MCPServerFormFields: React.FC<MCPServerFormFieldsProps> = ({
-  mode,
-  transport,
-  onTransportChange,
-  authType = 'none',
-  onAuthTypeChange,
-  form,
-  client,
-  serverId,
-  onTestConnection,
-  testing = false,
-  testResult,
-  onSaveFirst,
-}) => {
-  const { showSuccess, showError, showWarning, showInfo } = useThemedMessage();
-  const [testingAuth, setTestingAuth] = useState(false);
-  const [oauthBrowserFlowAvailable, setOauthBrowserFlowAvailable] = useState(false);
-  const [startingOAuthFlow, setStartingOAuthFlow] = useState(false);
-
-  // OAuth flow state
-  const [oauthCallbackModalVisible, setOauthCallbackModalVisible] = useState(false);
-  const [disconnectingOAuth, setDisconnectingOAuth] = useState(false);
-  const oauthCompletedCleanupRef = useRef<(() => void) | null>(null);
-
-  // Clean up OAuth listener on unmount
-  useEffect(() => {
-    return () => {
-      oauthCompletedCleanupRef.current?.();
-    };
-  }, []);
-
-  // Track effective server ID (may differ from prop after onSaveFirst creates a new server)
-  const [effectiveServerId, setEffectiveServerId] = useState<string | undefined>(serverId);
-  useEffect(() => {
-    setEffectiveServerId(serverId);
-  }, [serverId]);
-
-  // Start the browser-based OAuth 2.1 flow (two-phase for remote daemon)
-  const handleStartOAuthFlow = async () => {
-    if (!client) {
-      showError('Client not available');
-      return;
-    }
-
-    // Track the target server ID (may be set after saving)
-    let targetServerId = effectiveServerId;
-
-    // If no serverId and we have onSaveFirst callback, save the server first
-    if (!targetServerId && onSaveFirst) {
-      showInfo('Saving MCP server before testing...');
-      const newServerId = await onSaveFirst();
-      if (!newServerId) {
-        showError('Failed to save MCP server');
-        return;
-      }
-      targetServerId = newServerId;
-      setEffectiveServerId(newServerId);
-    }
-
-    const values = form.getFieldsValue(true);
-    const requestData = extractOAuthConfigForTesting(values);
-    if (!requestData) {
-      showError('MCP URL is required');
-      return;
-    }
-
-    setStartingOAuthFlow(true);
-
-    // Set up listener for oauth:open_browser event from daemon
-    const handleOpenBrowser = ({ authUrl }: { authUrl: string }) => {
-      console.log('[OAuth] Received open_browser event, opening:', authUrl);
-      window.open(authUrl, '_blank', 'noopener,noreferrer');
-    };
-    client.io.on('oauth:open_browser', handleOpenBrowser);
-
-    try {
-      showInfo('Starting OAuth authentication flow...');
-
-      // Use the new two-phase OAuth flow
-      const data = (await client.service('mcp-servers/oauth-start').create({
-        mcp_url: requestData.mcp_url,
-        mcp_server_id: targetServerId,
-        client_id: requestData.client_id,
-      })) as {
-        success: boolean;
-        error?: string;
-        message?: string;
-        authorizationUrl?: string;
-        state?: string;
-      };
-
-      if (data.success && data.state) {
-        setOauthCallbackModalVisible(true);
-        showInfo('Authenticating... complete sign-in in the new tab.');
-
-        // Listen for automatic completion via the daemon's callback endpoint
-        const handleOAuthCompleted = (event: { state: string; success: boolean }) => {
-          if (event.state === data.state && event.success) {
-            showSuccess('OAuth authentication successful!');
-            setOauthCallbackModalVisible(false);
-            setOauthBrowserFlowAvailable(false);
-            cleanup();
-          }
-        };
-        const cleanup = () => {
-          client.io.off('oauth:completed', handleOAuthCompleted);
-          oauthCompletedCleanupRef.current = null;
-        };
-        // Store cleanup so it can be called on modal cancel
-        oauthCompletedCleanupRef.current?.();
-        oauthCompletedCleanupRef.current = cleanup;
-        client.io.on('oauth:completed', handleOAuthCompleted);
-      } else {
-        showError(data.error || 'Failed to start OAuth flow');
-      }
-    } catch (error) {
-      showError(`OAuth flow error: ${error instanceof Error ? error.message : String(error)}`);
-    } finally {
-      client.io.off('oauth:open_browser', handleOpenBrowser);
-      setStartingOAuthFlow(false);
-    }
-  };
-
-  // Disconnect OAuth - remove stored tokens
-  const handleDisconnectOAuth = async () => {
-    if (!client) {
-      showError('Client not available');
-      return;
-    }
-
-    if (!effectiveServerId) {
-      showError('Cannot disconnect: MCP server must be saved first');
-      return;
-    }
-
-    setDisconnectingOAuth(true);
-    try {
-      const data = (await client.service('mcp-servers/oauth-disconnect').create({
-        mcp_server_id: effectiveServerId,
-      })) as { success: boolean; message?: string; error?: string };
-
-      if (data.success) {
-        showSuccess(data.message || 'OAuth connection removed');
-        // Show the "Start OAuth Flow" button again so user can re-authenticate
-        setOauthBrowserFlowAvailable(true);
-      } else {
-        showError(data.error || 'Failed to disconnect OAuth');
-      }
-    } catch (error) {
-      showError(`Disconnect error: ${error instanceof Error ? error.message : String(error)}`);
-    } finally {
-      setDisconnectingOAuth(false);
-    }
-  };
-
-  const handleTestAuth = async () => {
-    if (!client) {
-      showError('Client not available');
-      return;
-    }
-
-    const values = form.getFieldsValue(true);
-    const currentAuthType = values.auth_type || authType;
-
-    setTestingAuth(true);
-    try {
-      if (currentAuthType === 'jwt') {
-        const apiUrl = values.jwt_api_url;
-        const apiToken = values.jwt_api_token;
-        const apiSecret = values.jwt_api_secret;
-
-        if (!apiUrl || !apiToken || !apiSecret) {
-          showError('Please fill in all JWT authentication fields');
-          return;
-        }
-
-        // Use Feathers client for authenticated request
-        const data = (await client.service('mcp-servers/test-jwt').create({
-          api_url: apiUrl,
-          api_token: apiToken,
-          api_secret: apiSecret,
-        })) as { success: boolean; error?: string };
-
-        if (data.success) {
-          showSuccess('JWT authentication successful - token received');
-        } else {
-          showError(data.error || 'JWT authentication failed');
-        }
-      } else if (currentAuthType === 'oauth') {
-        const requestData = extractOAuthConfigForTesting(values);
-        if (!requestData) {
-          showWarning('Please enter MCP URL first to test OAuth authentication');
-          return;
-        }
-
-        const data = (await client.service('mcp-servers/test-oauth').create(requestData)) as {
-          success: boolean;
-          error?: string;
-          message?: string;
-          oauthType?: string;
-          tokenValid?: boolean;
-          mcpStatus?: number;
-          mcpStatusText?: string;
-          tokenUrlSource?: string;
-          requiresBrowserFlow?: boolean;
-          metadataUrl?: string;
-          authorizationServers?: string[];
-          wwwAuthenticate?: string;
-          responseHeaders?: Record<string, string>;
-          hint?: string;
-          debugInfo?: unknown;
-        };
-
-        if (data.success) {
-          if (data.requiresBrowserFlow) {
-            // OAuth 2.1 detected but needs browser authentication
-            setOauthBrowserFlowAvailable(true);
-            showInfo(
-              data.message ||
-                'OAuth 2.1 detected. Click "Start OAuth Flow" to authenticate in browser.'
-            );
-          } else if (data.oauthType === 'none') {
-            setOauthBrowserFlowAvailable(false);
-            showSuccess('MCP server accessible without authentication');
-          } else {
-            let message = data.message || 'OAuth authentication successful';
-            if (data.tokenUrlSource === 'auto-detected') {
-              message += ' (token URL auto-detected)';
-            }
-            if (data.mcpStatus !== undefined) {
-              message += ` | MCP server responded with ${data.mcpStatus}`;
-            }
-            showSuccess(message);
-          }
-        } else {
-          // Show detailed error with hints
-          let errorMsg = data.error || 'OAuth authentication failed';
-          if (data.hint) {
-            errorMsg += `\n\nHint: ${data.hint}`;
-          }
-          showError(errorMsg);
-        }
-      } else if (currentAuthType === 'bearer') {
-        const token = values.auth_token;
-        if (token) {
-          showSuccess('Bearer token configured');
-        } else {
-          showWarning('No bearer token provided');
-        }
-      } else {
-        // Auth type is 'none' - just confirm the configuration
-        showInfo('No authentication required - ready to use');
-      }
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-      showError(`Connection test failed: ${errorMessage}`);
-    } finally {
-      setTestingAuth(false);
-    }
-  };
-
-  const collapseItems = [
-    // Basic Info - always shown, not in collapse
-    ...(mode === 'create'
-      ? [
-          {
-            key: 'basic',
-            label: <Typography.Text strong>Basic Information</Typography.Text>,
-            children: (
-              <>
-                <Form.Item
-                  label="Name (Internal ID)"
-                  name="name"
-                  rules={[{ required: true, message: 'Please enter a server name' }]}
-                  tooltip="Internal identifier - lowercase, no spaces (e.g., filesystem, sentry, context7)"
-                >
-                  <Input placeholder="context7" />
-                </Form.Item>
-
-                <Form.Item
-                  label="Scope"
-                  name="scope"
-                  initialValue="session"
-                  tooltip="Where this server is available"
-                >
-                  <Select>
-                    <Select.Option value="global">Global (all sessions)</Select.Option>
-                    <Select.Option value="session">Session</Select.Option>
-                  </Select>
-                </Form.Item>
-
-                <Form.Item
-                  label="Display Name (Optional)"
-                  name="display_name"
-                  tooltip="User-friendly name shown in UI (e.g., Context7 MCP)"
-                >
-                  <Input placeholder="Context7 MCP" />
-                </Form.Item>
-
-                <Form.Item
-                  label="Enabled"
-                  name="enabled"
-                  valuePropName="checked"
-                  initialValue={true}
-                >
-                  <Switch />
-                </Form.Item>
-
-                <Form.Item label="Description" name="description">
-                  <TextArea placeholder="Optional description..." rows={2} />
-                </Form.Item>
-              </>
-            ),
-          },
-        ]
-      : [
-          {
-            key: 'basic',
-            label: <Typography.Text strong>Basic Information</Typography.Text>,
-            children: (
-              <>
-                <Form.Item
-                  label="Name (Internal ID)"
-                  name="name"
-                  tooltip="Internal identifier - cannot be changed after creation"
-                >
-                  <Input disabled />
-                </Form.Item>
-
-                <Form.Item
-                  label="Scope"
-                  name="scope"
-                  initialValue="global"
-                  tooltip="Where this server is available"
-                >
-                  <Select>
-                    <Select.Option value="global">Global (all sessions)</Select.Option>
-                    <Select.Option value="session">Session</Select.Option>
-                  </Select>
-                </Form.Item>
-
-                <Form.Item label="Display Name" name="display_name">
-                  <Input placeholder="Filesystem Access" />
-                </Form.Item>
-
-                <Form.Item
-                  label="Enabled"
-                  name="enabled"
-                  valuePropName="checked"
-                  initialValue={true}
-                >
-                  <Switch />
-                </Form.Item>
-
-                <Form.Item label="Description" name="description">
-                  <TextArea placeholder="Optional description..." rows={2} />
-                </Form.Item>
-              </>
-            ),
-          },
-        ]),
-    {
-      key: 'connection',
-      label: <Typography.Text strong>Connection</Typography.Text>,
-      children: (
-        <>
-          <Alert
-            title={
-              <>
-                Use <Typography.Text code>{'{{ user.env.VAR }}'}</Typography.Text> to inject your
-                environment variables.
-              </>
-            }
-            type="info"
-            showIcon
-            style={{ marginBottom: 16 }}
-          />
-
-          <Form.Item
-            label="Transport"
-            name="transport"
-            rules={mode === 'create' ? [{ required: true }] : []}
-            initialValue={mode === 'create' ? 'stdio' : undefined}
-            tooltip="Connection method: stdio for local processes, HTTP/SSE for remote servers"
-          >
-            <Select onChange={(value) => onTransportChange?.(value as 'stdio' | 'http' | 'sse')}>
-              <Select.Option value="stdio">stdio (Local process)</Select.Option>
-              <Select.Option value="http">HTTP</Select.Option>
-              <Select.Option value="sse">SSE (Server-Sent Events)</Select.Option>
-            </Select>
-          </Form.Item>
-
-          {transport === 'stdio' ? (
-            <>
-              <Form.Item
-                label="Command"
-                name="command"
-                rules={
-                  mode === 'create' ? [{ required: true, message: 'Please enter a command' }] : []
-                }
-                tooltip="Command to execute (e.g., npx, node, python)"
-              >
-                <Input placeholder="npx" />
-              </Form.Item>
-
-              <Form.Item
-                label="Arguments"
-                name="args"
-                tooltip="Comma-separated arguments. Each argument will be passed separately to the command. Example: -y, @modelcontextprotocol/server-filesystem, /allowed/path"
-              >
-                <Input placeholder="-y, @modelcontextprotocol/server-filesystem, /allowed/path" />
-              </Form.Item>
-            </>
-          ) : (
-            <>
-              <Form.Item
-                label="URL"
-                name="url"
-                rules={mode === 'create' ? [{ required: true, message: 'Please enter a URL' }] : []}
-                tooltip="Server URL. Supports templates like {{ user.env.MCP_URL }}"
-              >
-                <Input placeholder="https://mcp.example.com" />
-              </Form.Item>
-
-              <Form.Item
-                label="Auth Type"
-                name="auth_type"
-                initialValue="none"
-                tooltip="Authentication method for the MCP server"
-              >
-                <Select
-                  onChange={(value) => {
-                    // Reset OAuth flow state when changing auth type
-                    setOauthBrowserFlowAvailable(false);
-
-                    // Clear fields from other auth types to prevent validation issues
-                    if (value !== 'jwt') {
-                      form.setFieldsValue({
-                        jwt_api_url: undefined,
-                        jwt_api_token: undefined,
-                        jwt_api_secret: undefined,
-                      });
-                    }
-                    if (value !== 'bearer') {
-                      form.setFieldsValue({ auth_token: undefined });
-                    }
-                    if (value !== 'oauth') {
-                      form.setFieldsValue({
-                        oauth_authorization_url: undefined,
-                        oauth_token_url: undefined,
-                        oauth_client_id: undefined,
-                        oauth_client_secret: undefined,
-                        oauth_scope: undefined,
-                      });
-                    }
-                    onAuthTypeChange?.(value as 'none' | 'bearer' | 'jwt' | 'oauth');
-                  }}
-                >
-                  <Select.Option value="none">None</Select.Option>
-                  <Select.Option value="bearer">Bearer Token</Select.Option>
-                  <Select.Option value="jwt">JWT</Select.Option>
-                  <Select.Option value="oauth">OAuth 2.1</Select.Option>
-                </Select>
-              </Form.Item>
-
-              {authType === 'bearer' && (
-                <Form.Item
-                  label="Token"
-                  name="auth_token"
-                  rules={
-                    authType === 'bearer'
-                      ? [{ required: true, message: 'Please enter a bearer token' }]
-                      : []
-                  }
-                  tooltip="Bearer token. Supports templates like {{ user.env.API_TOKEN }}"
-                >
-                  <Input.Password placeholder="{{ user.env.API_TOKEN }} or raw token" />
-                </Form.Item>
-              )}
-
-              {authType === 'jwt' && (
-                <>
-                  <Form.Item
-                    label="API URL"
-                    name="jwt_api_url"
-                    rules={
-                      authType === 'jwt'
-                        ? [{ required: true, message: 'Please enter the API URL' }]
-                        : []
-                    }
-                    tooltip="JWT auth API URL. Supports templates."
-                  >
-                    <Input placeholder="https://auth.example.com/token" />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="API Token"
-                    name="jwt_api_token"
-                    rules={
-                      authType === 'jwt'
-                        ? [{ required: true, message: 'Please enter the API token' }]
-                        : []
-                    }
-                    tooltip="JWT API token. Supports templates like {{ user.env.JWT_TOKEN }}"
-                  >
-                    <Input.Password placeholder="{{ user.env.JWT_TOKEN }} or raw token" />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="API Secret"
-                    name="jwt_api_secret"
-                    rules={
-                      authType === 'jwt'
-                        ? [{ required: true, message: 'Please enter the API secret' }]
-                        : []
-                    }
-                    tooltip="JWT API secret. Supports templates like {{ user.env.JWT_SECRET }}"
-                  >
-                    <Input.Password placeholder="{{ user.env.JWT_SECRET }} or raw secret" />
-                  </Form.Item>
-                </>
-              )}
-
-              {authType === 'oauth' && (
-                <>
-                  <Form.Item
-                    label="Authorization URL"
-                    name="oauth_authorization_url"
-                    tooltip="OAuth authorization endpoint for browser-based login. Leave empty for auto-discovery (RFC 8414)."
-                  >
-                    <Input placeholder="https://auth.example.com/oauth/authorize" allowClear />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="Token URL"
-                    name="oauth_token_url"
-                    tooltip="OAuth token endpoint. Leave empty for auto-discovery (OAuth 2.1 RFC 9728)"
-                  >
-                    <Input placeholder="Auto-detect or {{ user.env.OAUTH_TOKEN_URL }}" allowClear />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="Client ID"
-                    name="oauth_client_id"
-                    tooltip="Required for servers that don't support Dynamic Client Registration (e.g. Figma, GitHub). Register an OAuth app with the provider and paste the client ID here."
-                  >
-                    <Input
-                      placeholder="Enter client ID or {{ user.env.OAUTH_CLIENT_ID }}"
-                      allowClear
-                    />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="Client Secret"
-                    name="oauth_client_secret"
-                    tooltip="Required for servers that use confidential clients (e.g. Figma). The secret is sent via HTTP Basic Auth during token exchange."
-                  >
-                    <Input.Password
-                      placeholder="Enter client secret or {{ user.env.OAUTH_CLIENT_SECRET }}"
-                      allowClear
-                    />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="Scope"
-                    name="oauth_scope"
-                    tooltip="Optional: OAuth scopes (space-separated, e.g., 'read write')"
-                  >
-                    <Input placeholder="Leave empty or specify scopes" allowClear />
-                  </Form.Item>
-
-                  <Form.Item
-                    label="Grant Type"
-                    name="oauth_grant_type"
-                    initialValue="client_credentials"
-                    tooltip="OAuth grant type for Client Credentials flow. OAuth 2.1 auto-discovery uses Authorization Code with PKCE instead."
-                  >
-                    <Select disabled>
-                      <Select.Option value="client_credentials">Client Credentials</Select.Option>
-                    </Select>
-                  </Form.Item>
-
-                  <Form.Item
-                    label="OAuth Mode"
-                    name="oauth_mode"
-                    initialValue="per_user"
-                    tooltip="Per User: Each user authenticates separately (recommended). Shared: One token for all users."
-                  >
-                    <Select>
-                      <Select.Option value="per_user">
-                        Per User (each user authenticates) - Recommended
-                      </Select.Option>
-                      <Select.Option value="shared">
-                        Shared (single token for all users)
-                      </Select.Option>
-                    </Select>
-                  </Form.Item>
-
-                  <Alert
-                    title="OAuth Configuration"
-                    description={
-                      <ul style={{ margin: 0, paddingLeft: 20, fontSize: 12 }}>
-                        <li>
-                          <strong>Auto-discovery:</strong> OAuth 2.1 servers advertise their
-                          endpoints automatically
-                        </li>
-                        <li>
-                          <strong>Pre-registered apps:</strong> Some servers (e.g. Figma, GitHub)
-                          require you to register an OAuth app in their developer portal and enter
-                          the Client ID and Client Secret here
-                        </li>
-                        <li>When Client ID is provided, Dynamic Client Registration is skipped</li>
-                        <li>Browser opens automatically for user authentication (PKCE flow)</li>
-                      </ul>
-                    }
-                    type="info"
-                    showIcon
-                    style={{ marginBottom: 16 }}
-                  />
-                </>
-              )}
-
-              {authType !== 'none' && (
-                <Form.Item>
-                  <Space>
-                    <Button type="default" loading={testingAuth} onClick={handleTestAuth}>
-                      Test Authentication
-                    </Button>
-                    {authType === 'oauth' && oauthBrowserFlowAvailable && (
-                      <Button
-                        type="primary"
-                        loading={startingOAuthFlow}
-                        onClick={handleStartOAuthFlow}
-                      >
-                        Start OAuth Flow
-                      </Button>
-                    )}
-                    {authType === 'oauth' && effectiveServerId && !oauthBrowserFlowAvailable && (
-                      <Button
-                        type="default"
-                        danger
-                        loading={disconnectingOAuth}
-                        onClick={handleDisconnectOAuth}
-                      >
-                        Disconnect OAuth
-                      </Button>
-                    )}
-                  </Space>
-                </Form.Item>
-              )}
-            </>
-          )}
-
-          <Form.Item
-            label="Environment Variables"
-            name="env"
-            tooltip="JSON object of environment variables. Values support templates like {{ user.env.VAR_NAME }}"
-          >
-            <TextArea
-              placeholder='{"GITHUB_TOKEN": "{{ user.env.GITHUB_TOKEN }}", "ALLOWED_PATHS": "/path"}'
-              rows={3}
-            />
-          </Form.Item>
-
-          {/* Test Connection - only for HTTP/SSE transport */}
-          {transport !== 'stdio' && (
-            <div style={{ borderTop: '1px solid #303030', marginTop: 16, paddingTop: 16 }}>
-              <Space orientation="vertical" style={{ width: '100%' }}>
-                <Button
-                  type="default"
-                  icon={<ApiOutlined />}
-                  onClick={onTestConnection}
-                  loading={testing}
-                >
-                  {testing ? 'Testing...' : 'Test Connection'}
-                </Button>
-
-                {testResult?.success && (
-                  <div style={{ marginTop: 8 }}>
-                    <Alert
-                      type="success"
-                      title={`Connected: ${testResult.toolCount} tools, ${testResult.resourceCount} resources, ${testResult.promptCount} prompts`}
-                      showIcon
-                      style={{ marginBottom: 8 }}
-                    />
-                    {testResult.tools && testResult.tools.length > 0 && (
-                      <div style={{ marginTop: 8 }}>
-                        <Typography.Text
-                          type="secondary"
-                          style={{ fontSize: 12, display: 'block', marginBottom: 4 }}
-                        >
-                          Tools:
-                        </Typography.Text>
-                        <div style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>
-                          {testResult.tools.map((tool) => (
-                            <Tooltip
-                              key={tool.name}
-                              title={tool.description || 'No description'}
-                              placement="top"
-                            >
-                              <Tag color="blue" style={{ marginBottom: 4, cursor: 'help' }}>
-                                {tool.name}
-                              </Tag>
-                            </Tooltip>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                    {testResult.resources && testResult.resources.length > 0 && (
-                      <div style={{ marginTop: 8 }}>
-                        <Typography.Text
-                          type="secondary"
-                          style={{ fontSize: 12, display: 'block', marginBottom: 4 }}
-                        >
-                          Resources:
-                        </Typography.Text>
-                        <div style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>
-                          {testResult.resources.map((resource) => (
-                            <Tooltip
-                              key={resource.uri}
-                              title={
-                                <div>
-                                  <div>{resource.uri}</div>
-                                  {resource.mimeType && (
-                                    <div style={{ opacity: 0.7, fontSize: 11 }}>
-                                      {resource.mimeType}
-                                    </div>
-                                  )}
-                                </div>
-                              }
-                              placement="top"
-                            >
-                              <Tag color="cyan" style={{ marginBottom: 4, cursor: 'help' }}>
-                                {resource.name}
-                              </Tag>
-                            </Tooltip>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                    {testResult.prompts && testResult.prompts.length > 0 && (
-                      <div style={{ marginTop: 8 }}>
-                        <Typography.Text
-                          type="secondary"
-                          style={{ fontSize: 12, display: 'block', marginBottom: 4 }}
-                        >
-                          Prompts:
-                        </Typography.Text>
-                        <div style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>
-                          {testResult.prompts.map((prompt) => (
-                            <Tooltip
-                              key={prompt.name}
-                              title={prompt.description || 'No description'}
-                              placement="top"
-                            >
-                              <Tag color="purple" style={{ marginBottom: 4, cursor: 'help' }}>
-                                {prompt.name}
-                              </Tag>
-                            </Tooltip>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                )}
-
-                {testResult && !testResult.success && (
-                  <Alert
-                    type="error"
-                    title="Connection failed"
-                    description={testResult.error}
-                    showIcon
-                    style={{ marginTop: 8 }}
-                  />
-                )}
-              </Space>
-            </div>
-          )}
-        </>
-      ),
-    },
-  ];
-
-  return (
-    <>
-      <Collapse
-        ghost
-        destroyOnHidden={false}
-        defaultActiveKey={['basic']}
-        expandIcon={({ isActive }) => <DownOutlined rotate={isActive ? 180 : 0} />}
-        items={collapseItems}
-      />
-
-      {/* OAuth waiting modal - closes automatically when daemon receives the callback */}
-      <Modal
-        title="OAuth Authentication"
-        open={oauthCallbackModalVisible}
-        onCancel={() => {
-          setOauthCallbackModalVisible(false);
-          oauthCompletedCleanupRef.current?.();
-        }}
-        footer={[
-          <Button
-            key="cancel"
-            onClick={() => {
-              setOauthCallbackModalVisible(false);
-              oauthCompletedCleanupRef.current?.();
-            }}
-          >
-            Cancel
-          </Button>,
-        ]}
-      >
-        <Space orientation="vertical" style={{ width: '100%' }}>
-          <Typography.Paragraph>
-            Waiting for authentication to complete in the browser tab...
-          </Typography.Paragraph>
-          <Typography.Paragraph>
-            This dialog will close automatically once sign-in is complete.
-          </Typography.Paragraph>
-        </Space>
-      </Modal>
-    </>
-  );
-};
 
 export const MCPServersTable: React.FC<MCPServersTableProps> = ({
   mcpServerById,
@@ -905,24 +51,15 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
   const [viewModalOpen, setViewModalOpen] = useState(false);
   const [editingServer, setEditingServer] = useState<MCPServer | null>(null);
   const [viewingServer, setViewingServer] = useState<MCPServer | null>(null);
-  const [form] = Form.useForm();
+  const [createForm] = Form.useForm();
   const [transport, setTransport] = useState<'stdio' | 'http' | 'sse'>('stdio');
   const [authType, setAuthType] = useState<'none' | 'bearer' | 'jwt' | 'oauth'>('none');
   const [testing, setTesting] = useState(false);
-  // Track if server was already created by onSaveFirst during OAuth flow
   const [alreadyCreatedInOAuthFlow, setAlreadyCreatedInOAuthFlow] = useState(false);
-  const [testResult, setTestResult] = useState<{
-    success: boolean;
-    toolCount: number;
-    resourceCount: number;
-    promptCount: number;
-    error?: string;
-    tools?: Array<{ name: string; description: string }>;
-    resources?: Array<{ name: string; uri: string; mimeType?: string }>;
-    prompts?: Array<{ name: string; description: string }>;
-  } | null>(null);
+  const [testResult, setTestResult] = useState<TestResult | null>(null);
 
-  // Sync editing server when mcpServerById updates (real-time WebSocket updates)
+  // Sync editing server when mcpServerById updates (real-time WebSocket updates).
+  // Also keeps the open edit modal in sync if the underlying record changes.
   useEffect(() => {
     if (editingServer && mcpServerById.has(editingServer.mcp_server_id)) {
       const updatedServer = mcpServerById.get(editingServer.mcp_server_id);
@@ -932,51 +69,54 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
     }
   }, [mcpServerById, editingServer]);
 
+  const buildCreateData = (values: Record<string, unknown>): CreateMCPServerInput => {
+    const data: CreateMCPServerInput = {
+      name: values.name as string,
+      display_name: values.display_name as string | undefined,
+      description: values.description as string | undefined,
+      transport: values.transport as 'stdio' | 'http' | 'sse',
+      scope: (values.scope as 'global' | 'session' | undefined) || 'global',
+      enabled: (values.enabled as boolean | undefined) ?? true,
+      source: 'user',
+    };
+
+    if (values.transport === 'stdio') {
+      data.command = values.command as string;
+      data.args = (values.args as string)?.split(',').map((arg: string) => arg.trim()) || [];
+    } else {
+      data.url = values.url as string;
+    }
+
+    if (values.auth_type && values.auth_type !== 'none') {
+      data.auth = { type: values.auth_type as 'bearer' | 'jwt' | 'oauth' };
+      if (values.auth_type === 'bearer') {
+        data.auth.token = values.auth_token as string;
+      } else if (values.auth_type === 'jwt') {
+        data.auth.api_url = values.jwt_api_url as string;
+        data.auth.api_token = values.jwt_api_token as string;
+        data.auth.api_secret = values.jwt_api_secret as string;
+      } else if (values.auth_type === 'oauth') {
+        Object.assign(data.auth, extractOAuthConfig(values));
+      }
+    }
+
+    if (values.env) {
+      try {
+        data.env = JSON.parse(values.env as string);
+      } catch {
+        // Invalid JSON, skip
+      }
+    }
+
+    return data;
+  };
+
   // Save server first for OAuth flow in create mode (returns new server ID)
   const handleSaveFirstForCreate = async (): Promise<string | null> => {
     if (!client) return null;
     try {
-      await form.validateFields();
-      const values = form.getFieldsValue(true);
-      const data: CreateMCPServerInput = {
-        name: values.name,
-        display_name: values.display_name,
-        description: values.description,
-        transport: values.transport,
-        scope: values.scope || 'global',
-        enabled: values.enabled ?? true,
-        source: 'user',
-      };
-
-      if (values.transport === 'stdio') {
-        data.command = values.command;
-        data.args = values.args?.split(',').map((arg: string) => arg.trim()) || [];
-      } else {
-        data.url = values.url;
-      }
-
-      if (values.auth_type && values.auth_type !== 'none') {
-        data.auth = { type: values.auth_type };
-        if (values.auth_type === 'bearer') {
-          data.auth.token = values.auth_token;
-        } else if (values.auth_type === 'jwt') {
-          data.auth.api_url = values.jwt_api_url;
-          data.auth.api_token = values.jwt_api_token;
-          data.auth.api_secret = values.jwt_api_secret;
-        } else if (values.auth_type === 'oauth') {
-          const oauthConfig = extractOAuthConfig(values);
-          Object.assign(data.auth, oauthConfig);
-        }
-      }
-
-      if (values.env) {
-        try {
-          data.env = JSON.parse(values.env);
-        } catch {
-          // Invalid JSON, skip
-        }
-      }
-
+      await createForm.validateFields();
+      const data = buildCreateData(createForm.getFieldsValue(true));
       const result = await client.service('mcp-servers').create(data);
       setAlreadyCreatedInOAuthFlow(true);
       return (result as MCPServer).mcp_server_id || null;
@@ -985,78 +125,30 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
     }
   };
 
+  const resetCreateModal = () => {
+    createForm.resetFields();
+    setCreateModalOpen(false);
+    setTransport('stdio');
+    setAuthType('none');
+    setTestResult(null);
+    setAlreadyCreatedInOAuthFlow(false);
+  };
+
   const handleCreate = () => {
-    // If server was already created during OAuth flow, just close the modal
     if (alreadyCreatedInOAuthFlow) {
-      form.resetFields();
-      setCreateModalOpen(false);
-      setTransport('stdio');
-      setAuthType('none');
-      setTestResult(null);
-      setAlreadyCreatedInOAuthFlow(false);
+      resetCreateModal();
       return;
     }
 
-    form
+    createForm
       .validateFields()
       .then(() => {
-        const values = form.getFieldsValue(true);
-        const data: CreateMCPServerInput = {
-          name: values.name,
-          display_name: values.display_name,
-          description: values.description,
-          transport: values.transport,
-          scope: values.scope || 'global',
-          enabled: values.enabled ?? true,
-          source: 'user',
-        };
-
-        // Add transport-specific fields
-        if (values.transport === 'stdio') {
-          data.command = values.command;
-          data.args = values.args?.split(',').map((arg: string) => arg.trim()) || [];
-        } else {
-          data.url = values.url;
-        }
-
-        // Add auth config if present
-        if (values.auth_type && values.auth_type !== 'none') {
-          data.auth = {
-            type: values.auth_type,
-          };
-          if (values.auth_type === 'bearer') {
-            data.auth.token = values.auth_token;
-          } else if (values.auth_type === 'jwt') {
-            data.auth.api_url = values.jwt_api_url;
-            data.auth.api_token = values.jwt_api_token;
-            data.auth.api_secret = values.jwt_api_secret;
-          } else if (values.auth_type === 'oauth') {
-            // Extract OAuth config using helper (only includes non-empty fields)
-            const oauthConfig = extractOAuthConfig(values);
-            Object.assign(data.auth, oauthConfig);
-          }
-        }
-
-        // Add env vars if present
-        if (values.env) {
-          try {
-            data.env = JSON.parse(values.env);
-          } catch {
-            // Invalid JSON, skip
-          }
-        }
-
+        const data = buildCreateData(createForm.getFieldsValue(true));
         onCreate?.(data);
-        form.resetFields();
-        setCreateModalOpen(false);
-        setTransport('stdio');
-        setAuthType('none');
-        setTestResult(null);
+        resetCreateModal();
       })
       .catch((error) => {
-        // Validation failed - form will show errors automatically
         console.error('Form validation failed:', error);
-        // Error fields are shown inline by Ant Design Form
         if (error.errorFields && error.errorFields.length > 0) {
           const firstError = error.errorFields[0];
           showError(firstError.errors[0] || 'Please fill in required fields');
@@ -1064,22 +156,19 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
       });
   };
 
-  // Test connection using current form values (inline config, no save required)
-  // If serverId is provided, capabilities will be persisted after successful test
-  const handleTestConnection = async (serverId?: string) => {
+  // Test connection from create modal (always inline config, no persistence).
+  const handleCreateTestConnection = async () => {
     if (!client) {
       showError('Client not available');
       return;
     }
 
-    const values = form.getFieldsValue(true);
+    const values = createForm.getFieldsValue(true);
 
-    // Validate required fields for connection test
     if (!values.url) {
       showError('URL is required to test connection');
       return;
     }
-
     if (values.transport === 'stdio') {
       showError('Connection test is not available for stdio transport');
       return;
@@ -1089,7 +178,6 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
     setTestResult(null);
 
     try {
-      // Build auth config from form values
       let auth:
         | {
             type: 'none' | 'bearer' | 'jwt' | 'oauth';
@@ -1115,30 +203,15 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
           auth.api_token = values.jwt_api_token;
           auth.api_secret = values.jwt_api_secret;
         } else if (values.auth_type === 'oauth') {
-          // Extract OAuth config using helper
-          const oauthConfig = extractOAuthConfig(values);
-          Object.assign(auth, oauthConfig);
+          Object.assign(auth, extractOAuthConfig(values));
         }
       }
 
-      // Send form values for testing using inline config
-      const requestData: {
-        mcp_server_id?: string;
-        url: string;
-        transport: 'http' | 'sse';
-        auth?: typeof auth;
-      } = {
+      const data = (await client.service('mcp-servers/discover').create({
         url: values.url,
         transport: values.transport || 'http',
         auth,
-      };
-
-      // Include server ID to persist discovered capabilities (only for saved servers)
-      if (serverId) {
-        requestData.mcp_server_id = serverId;
-      }
-
-      const data = (await client.service('mcp-servers/discover').create(requestData)) as {
+      })) as {
         success: boolean;
         error?: string;
         capabilities?: { tools: number; resources: number; prompts: number };
@@ -1148,7 +221,7 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
       };
 
       if (data.success && data.capabilities) {
-        const result = {
+        const result: TestResult = {
           success: true,
           toolCount: data.capabilities.tools,
           resourceCount: data.capabilities.resources,
@@ -1174,7 +247,6 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
       }
     } catch (error) {
       const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-      console.error('Connection test failed:', error);
       setTestResult({
         success: false,
         toolCount: 0,
@@ -1188,131 +260,14 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
     }
   };
 
-  const handleEdit = async (server: MCPServer) => {
+  const handleEdit = (server: MCPServer) => {
     setEditingServer(server);
-    setTestResult(null); // Reset test result when opening edit modal
-    const serverAuthType = (server.auth?.type as 'none' | 'bearer' | 'jwt' | 'oauth') || 'none';
-    setAuthType(serverAuthType);
-
-    // Set transport state for conditional rendering
-    setTransport(server.transport || (server.url ? 'http' : 'stdio'));
-
-    // Reset form first to clear any stale registered fields from previous edits
-    // This prevents Ant Design from validating hidden fields (e.g. JWT fields when auth is OAuth)
-    form.resetFields();
-
-    // Set form fields - only include auth fields relevant to the current auth type
-    const formValues: Record<string, unknown> = {
-      name: server.name,
-      display_name: server.display_name,
-      description: server.description,
-      transport: server.transport || (server.url ? 'http' : 'stdio'),
-      command: server.command,
-      args: server.args?.join(', '),
-      url: server.url,
-      scope: server.scope,
-      enabled: server.enabled,
-      env: server.env ? JSON.stringify(server.env, null, 2) : undefined,
-      auth_type: serverAuthType,
-      tool_permissions: server.tool_permissions || {},
-    };
-
-    // Only set auth fields for the active auth type to avoid
-    // Ant Design registering and validating hidden fields
-    if (serverAuthType === 'bearer') {
-      formValues.auth_token = server.auth?.token;
-    } else if (serverAuthType === 'jwt') {
-      formValues.jwt_api_url = server.auth?.api_url;
-      formValues.jwt_api_token = server.auth?.api_token;
-      formValues.jwt_api_secret = server.auth?.api_secret;
-    } else if (serverAuthType === 'oauth') {
-      formValues.oauth_authorization_url = server.auth?.oauth_authorization_url;
-      formValues.oauth_token_url = server.auth?.oauth_token_url;
-      formValues.oauth_client_id = server.auth?.oauth_client_id;
-      formValues.oauth_client_secret = server.auth?.oauth_client_secret;
-      formValues.oauth_scope = server.auth?.oauth_scope;
-      formValues.oauth_grant_type = server.auth?.oauth_grant_type || 'client_credentials';
-      formValues.oauth_mode = server.auth?.oauth_mode || 'per_user';
-    }
-
-    form.setFieldsValue(formValues);
-
     setEditModalOpen(true);
   };
 
-  const handleUpdate = async () => {
-    if (!editingServer || !client) return;
-
-    try {
-      await form.validateFields();
-      const values = form.getFieldsValue(true);
-
-      const updates: UpdateMCPServerInput = {
-        display_name: values.display_name,
-        description: values.description,
-        scope: values.scope,
-        enabled: values.enabled,
-        transport: values.transport,
-      };
-
-      // Add transport-specific fields based on the NEW transport value
-      if (values.transport === 'stdio') {
-        updates.command = values.command;
-        updates.args = values.args?.split(',').map((arg: string) => arg.trim()) || [];
-      } else {
-        updates.url = values.url;
-      }
-
-      // Add env vars if present
-      if (values.env) {
-        try {
-          updates.env = JSON.parse(values.env);
-        } catch {
-          // Invalid JSON, skip
-        }
-      }
-
-      // Add auth config if present
-      if (values.auth_type && values.auth_type !== 'none') {
-        updates.auth = {
-          type: values.auth_type,
-        };
-        if (values.auth_type === 'bearer') {
-          updates.auth.token = values.auth_token;
-        } else if (values.auth_type === 'jwt') {
-          updates.auth.api_url = values.jwt_api_url;
-          updates.auth.api_token = values.jwt_api_token;
-          updates.auth.api_secret = values.jwt_api_secret;
-        } else if (values.auth_type === 'oauth') {
-          // Extract OAuth config using helper (only includes non-empty fields)
-          const oauthConfig = extractOAuthConfig(values);
-          Object.assign(updates.auth, oauthConfig);
-        }
-      } else {
-        updates.auth = undefined;
-      }
-
-      // Save the updates
-      await client.service('mcp-servers').patch(editingServer.mcp_server_id, updates);
-
-      // Also call parent callback for state management
-      onUpdate?.(editingServer.mcp_server_id, updates);
-
-      showSuccess('MCP server updated successfully');
-
-      // Close the modal after successful update
-      form.resetFields();
-      setEditModalOpen(false);
-      setEditingServer(null);
-      setTransport('stdio');
-      setAuthType('none');
-      setTestResult(null);
-    } catch (error) {
-      // Validation or update failed
-      console.error('Update failed:', error);
-      const errorMessage = error instanceof Error ? error.message : 'Failed to update server';
-      showError(errorMessage);
-    }
+  const handleEditClose = () => {
+    setEditModalOpen(false);
+    setEditingServer(null);
   };
 
   const handleView = (server: MCPServer) => {
@@ -1328,7 +283,6 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
     const toolCount = server.tools?.length || 0;
     const transport = server.transport || (server.url ? 'http' : 'stdio');
 
-    // For stdio servers, tools are only available when session is running
     if (transport === 'stdio') {
       return {
         status: 'default' as const,
@@ -1337,7 +291,6 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
       };
     }
 
-    // For HTTP/SSE servers, show tool count if tested
     if (toolCount > 0) {
       return {
         status: 'success' as const,
@@ -1494,27 +447,20 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
         title="Add MCP Server"
         open={createModalOpen}
         onOk={handleCreate}
-        onCancel={() => {
-          form.resetFields();
-          setCreateModalOpen(false);
-          setTransport('stdio');
-          setAuthType('none');
-          setTestResult(null);
-          setAlreadyCreatedInOAuthFlow(false);
-        }}
+        onCancel={resetCreateModal}
         okText={alreadyCreatedInOAuthFlow ? 'Done' : 'Create'}
         width={600}
       >
-        <Form form={form} layout="vertical" style={{ marginTop: 16 }}>
+        <Form form={createForm} layout="vertical" style={{ marginTop: 16 }}>
           <MCPServerFormFields
             mode="create"
             transport={transport}
             onTransportChange={setTransport}
             authType={authType}
             onAuthTypeChange={setAuthType}
-            form={form}
+            form={createForm}
             client={client}
-            onTestConnection={() => handleTestConnection()}
+            onTestConnection={handleCreateTestConnection}
             testing={testing}
             testResult={testResult}
             onSaveFirst={handleSaveFirstForCreate}
@@ -1522,40 +468,14 @@ export const MCPServersTable: React.FC<MCPServersTableProps> = ({
         </Form>
       </Modal>
 
-      {/* Edit MCP Server Modal */}
-      <Modal
-        title="Edit MCP Server"
+      {/* Edit MCP Server Modal — self-contained */}
+      <MCPServerEditModal
+        server={editingServer}
         open={editModalOpen}
-        onOk={handleUpdate}
-        onCancel={() => {
-          form.resetFields();
-          setEditModalOpen(false);
-          setEditingServer(null);
-          setTransport('stdio');
-          setAuthType('none');
-          setTestResult(null);
-        }}
-        okText="Save"
-        width={600}
-      >
-        <Form form={form} layout="vertical" style={{ marginTop: 16 }}>
-          <MCPServerFormFields
-            mode="edit"
-            transport={transport}
-            onTransportChange={setTransport}
-            authType={authType}
-            onAuthTypeChange={setAuthType}
-            form={form}
-            client={client}
-            serverId={editingServer?.mcp_server_id}
-            onTestConnection={
-              editingServer ? () => handleTestConnection(editingServer.mcp_server_id) : undefined
-            }
-            testing={testing}
-            testResult={testResult}
-          />
-        </Form>
-      </Modal>
+        client={client}
+        onClose={handleEditClose}
+        onUpdate={onUpdate}
+      />
 
       {/* View MCP Server Modal */}
       <Modal
